+++ src/pptlint/policy.py
@@ -1,0 +1,1 @@
+from decklint.policy import *  # noqa: F403